contact_management_service

+++ src/main/java/africa/semicolon/contactManagementService/services/groupService/GroupService.java
@@ -22,6 +22,6 @@ public interface GroupService {
 
     void removeContactFromGroup(RemoveContactFromGroupRequest request);
 
-    List<Contact> getContactsInGroup(Long groupId);
+    List<Contact> getContactsInGroup(int groupId);
     List<Group> searchGroups(SearchGroupsRequest request);
 }
